+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Update `smart` parameter to `kind` for inventory methods.
 - Fix `inventoryModify` method by adding mandatory `id` parameter.
 - Add `credentials` and `executionEnv` parameters to `jobTemplateLaunch` method.
+- Add `limit` and `skipTags` parameters to `workflowJobTemplateLaunch` method.
 
 **GoSS**
 - Add `logLevel` parameter to `server` and `validate` methods.